requirements.txt
langchain==0.0.161
python-dotenv==1.0.0
